Align object_file_system implementation with core
- Parts of core expect get_remote_file_path to represent all
of the file pool, where as previously used thisused the remote keyword
to signify a file was external, e.g. in S3. This was found to cause an
issue with a function in stored_file.
- To overcome this issue our remote functions now have
the same functionality as the old object functions. The old
remote functions have been renamed to external function