ObjectFS: Send correct expiry time for redirect URLs #422
When redirecting the user to download a file directly from S3, the
redirect was previously sent with whatever cache expiry time that
Moodle thought was appropriate for the file. In cases where this was
longer than a week, it caused problems, because S3 signed URLs can
only be valid for up to a week. This change sets the expiry on the
redirect to the same time that the URL expires at.