Removed legacy cron function and deprecated cron_trace_time_and_memory
In Moodle 4.2, cron_trace_time_and_memory has been replaced by a
different function in a namespaced location, and you get deprecation
warnings for including cronlib.php as well as for calling the function.

This change removes the old cron function in lib.php, and changes
the code that calls cron_trace_time_and_memory so that it uses the
old method prior to 4.2, and the new one after.